Add funding-aware event-driven backtesting engine
