+++ src/tradingbot/backtesting/__init__.py
@@ -1,0 +1,1 @@
+"""Backtesting utilities including event-driven engine."""
